third_party/blink/web_tests/external/wpt/scroll-animations/view-timelines/block-view-timeline-nested-subject.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] View timeline with subject that is not a direct descendant of the scroll container
  assert_equals: Effect is in the after phase at effect end time expected "0.7" but got "1"
Harness: the test ran to completion.
